card selector SelectedList converted to projectCardModel
