+++ front/.angular/cache/17.3.7/vite/deps/@angular_core.js
@@ -451,9 +451,8 @@ import {
   ɵɵvalidateIframeAttribute,
   ɵɵviewQuery,
   ɵɵviewQuerySignal
-} from "./chunk-I7GCJ5UE.js";
-import "./chunk-IBZ4ENFR.js";
-import "./chunk-3EWR26GJ.js";
+} from "./chunk-65CGMHPD.js";
+import "./chunk-VDFOJE7E.js";
 export {
   ANIMATION_MODULE_TYPE,
   APP_BOOTSTRAP_LISTENER,
