+++ front/.angular/cache/17.3.7/vite/deps/@angular_platform-browser.js
@@ -35,13 +35,12 @@ import {
   provideProtractorTestingSupport,
   withHttpTransferCacheOptions,
   withNoHttpTransferCache
-} from "./chunk-IUZGN5HH.js";
+} from "./chunk-IJDG6TYH.js";
 import {
   getDOM
-} from "./chunk-USCDVPWZ.js";
-import "./chunk-I7GCJ5UE.js";
-import "./chunk-IBZ4ENFR.js";
-import "./chunk-3EWR26GJ.js";
+} from "./chunk-Y3O3PVB6.js";
+import "./chunk-65CGMHPD.js";
+import "./chunk-VDFOJE7E.js";
 export {
   BrowserModule,
   By,
