+++ front/.angular/cache/17.3.7/vite/deps/@angular_router.js
@@ -1,6 +1,6 @@
 import {
   Title
-} from "./chunk-IUZGN5HH.js";
+} from "./chunk-IJDG6TYH.js";
 import {
   DOCUMENT,
   HashLocationStrategy,
@@ -9,7 +9,7 @@ import {
   LocationStrategy,
   PathLocationStrategy,
   ViewportScroller
-} from "./chunk-USCDVPWZ.js";
+} from "./chunk-Y3O3PVB6.js";
 import {
   APP_BOOTSTRAP_LISTENER,
   APP_INITIALIZER,
@@ -77,7 +77,7 @@ import {
   ɵɵloadQuery,
   ɵɵqueryRefresh,
   ɵɵsanitizeUrlOrResourceUrl
-} from "./chunk-I7GCJ5UE.js";
+} from "./chunk-65CGMHPD.js";
 import {
   BehaviorSubject,
   ConnectableObservable,
@@ -85,6 +85,9 @@ import {
   EmptyError,
   Subject,
   Subscription,
+  __async,
+  __spreadProps,
+  __spreadValues,
   catchError,
   combineLatest,
   concat,
@@ -112,12 +115,7 @@ import {
   takeUntil,
   tap,
   throwError
-} from "./chunk-IBZ4ENFR.js";
-import {
-  __async,
-  __spreadProps,
-  __spreadValues
-} from "./chunk-3EWR26GJ.js";
+} from "./chunk-VDFOJE7E.js";
 
 // node_modules/@angular/router/fesm2022/router.mjs
 var PRIMARY_OUTLET = "primary";
