--- front/.angular/cache/17.3.7/vite/deps/tslib.js
@@ -1,69 +1,0 @@
-import {
-  __addDisposableResource,
-  __assign,
-  __asyncDelegator,
-  __asyncGenerator,
-  __asyncValues,
-  __await,
-  __awaiter,
-  __classPrivateFieldGet,
-  __classPrivateFieldIn,
-  __classPrivateFieldSet,
-  __createBinding,
-  __decorate,
-  __disposeResources,
-  __esDecorate,
-  __exportStar,
-  __extends,
-  __generator,
-  __importDefault,
-  __importStar,
-  __makeTemplateObject,
-  __metadata,
-  __param,
-  __propKey,
-  __read,
-  __rest,
-  __runInitializers,
-  __setFunctionName,
-  __spread,
-  __spreadArray,
-  __spreadArrays,
-  __values,
-  tslib_es6_default
-} from "./chunk-3EWR26GJ.js";
-export {
-  __addDisposableResource,
-  __assign,
-  __asyncDelegator,
-  __asyncGenerator,
-  __asyncValues,
-  __await,
-  __awaiter,
-  __classPrivateFieldGet,
-  __classPrivateFieldIn,
-  __classPrivateFieldSet,
-  __createBinding,
-  __decorate,
-  __disposeResources,
-  __esDecorate,
-  __exportStar,
-  __extends,
-  __generator,
-  __importDefault,
-  __importStar,
-  __makeTemplateObject,
-  __metadata,
-  __param,
-  __propKey,
-  __read,
-  __rest,
-  __runInitializers,
-  __setFunctionName,
-  __spread,
-  __spreadArray,
-  __spreadArrays,
-  __values,
-  tslib_es6_default as default
-};
-//# sourceMappingURL=tslib.js.map
